:knife: Add compat to image_push for old Skopeo
TL;DR
=====
- An updated version of Skopeo isn't always available; ¯\_(ツ)_/¯
- Allow `image_push` to go into compatibility mode and do `copy`
  operations instead of `sync` if the image is less than v0.1.41

Detail
======
We wanted to be able to **still** push images to a registry _if_ the
version of Skopeo wasn't new enough to contain the `sync` feature. This,
by no means, is a replacement for `sync`. In fact, it is brittle, as it
isn't comparable to the routine that `sync` uses. More than anything,
this is something to unblock someone that may be in the predicament of
not being able to pull down an updated version of Skopeo.

Signed-off-by: Conzetti Finocchiaro <conzetti@gmail.com>