:bug: Provide Helm v3 support
TL;DR
=====
- Recent changes to the Helm CLI break this wrapper
- Only allow helm versions > `3.7.0`

Detail
======
The `chart_pull` functionality was recently crippled. Helm CLI changes
removed the `helm chart` command, so we're forced to leverage the
`helm pull` command to pull images. Since the TAC repository is
an OCI-compliant image registry, we need to _instead_ pull the image
using the new OCI format.

For example, our pull command would look like the following:
```
helm pull registry.pivotal.io/tac-federal-apple/charts/kubeapps --version 7.6.1
```

We are also doing version checks on the Helm CLI now, so anything that
is older than `3.7.0` will not work. This is a **breaking** change, so
use an older release if this is problematic for you.